renaiss-bot-py/requirements.txt
python-telegram-bot[ext]
python-dotenv
openai
aiohttp
sqlalchemy[asyncio]
aiosqlite
apscheduler
loguru
